Add unitary tests for BDS 0,0; 2,0; 4,0 and 5,0 Fix computation of negative values

+++ pkg/bds/bds50/fields/true_air_speed.go
@@ -2,13 +2,13 @@ package fields
 
 import "github.com/twuillemin/modes/pkg/bitutils"
 
-func ReadTrueAirSpeed(data []byte) (bool, float32) {
+func ReadTrueAirSpeed(data []byte) (bool, uint32) {
 	status := (data[5] & 0x04) != 0
 
 	byte1 := data[5] & 0x03
 	byte2 := data[6] & 0xFF
 	allBits := bitutils.Pack2Bytes(byte1, byte2)
-	speed := float32(allBits) * 2
+	speed := uint32(allBits) * 2
 
 	return status, speed
 }
